Whole project
This is an university project between the course Software Engineering.
Briefly what is it about. This project name is "cooperation". It's about
different firms what can  suggest their services to university for
example. We are storing data about those firms, about details of their
work, about their contact person, about persons who lead the projects
and so on. The main goal of this project was to create a GUI application
in Java which allows us to easily see, add, edit and delete data from
databaze. Since I am studying in Czech university the whole project is
written in Czech language, even the name of the methods and classes.